+++ core/src/main/java/com/draagon/meta/loader/file/FileMetaDataLoader.java
@@ -1,6 +1,5 @@
 package com.draagon.meta.loader.file;
 
-import com.draagon.meta.MetaData;
 import com.draagon.meta.loader.MetaDataLoader;
 import com.draagon.meta.loader.file.json.JsonMetaDataParser;
 import com.draagon.meta.loader.file.xml.XMLMetaDataParser;