+++ metadata/src/main/java/com/draagon/meta/enhancement/enhancers/TemplateBasedEnhancer.java
@@ -13,8 +13,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.Set;
+
 
 /**
  * v6.0.0: Enhancer that applies attributes based on template requirements.