+++ metadata/src/test/java/com/draagon/meta/io/object/gson/GsonAdapterTest.java
@@ -25,6 +25,7 @@ public class GsonAdapterTest {
         loader = SimpleLoader.createManual("proxytest", Arrays.asList(
                 "com/draagon/meta/loader/simple/fruitbasket-proxy-metadata.json"
         ));
+        
 
         builder = MetaObjectGsonInitializer.getBuilderWithAdapters(loader);
     }