+++ metaobjects-codegen/src/test/java/com/draagon/meta/generator/GeneratorTestBase.java
@@ -1,6 +1,10 @@
 package com.draagon.meta.generator;
 
+import com.draagon.meta.loader.simple.SimpleLoader;
 import com.draagon.meta.loader.simple.SimpleLoaderTestBase;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
 
 public class GeneratorTestBase extends SimpleLoaderTestBase {